requirements.txt
polars
fastapi
uvicorn[standard]
pydantic
requests
scipy
numpy
psycopg2-binary
